+++ lib/lrl/LRL_main.c
@@ -76,7 +76,7 @@ LRL_FileWriter *LRL_open_write_file(const char *filename)
  * \return null if failure and set status to error flag
  */
 LRL_RecordReader *LRL_open_read_record(LRL_FileReader *fr,
-				       size_t *rec_size, 
+				       off_t *rec_size, 
 				       LIME_type *lime_type, int *status)
 {
   LRL_RecordReader *rr;
@@ -124,7 +124,7 @@ LRL_RecordReader *LRL_open_read_record(LRL_FileReader *fr,
  */
 LRL_RecordWriter *LRL_open_write_record(LRL_FileWriter *fw, 
 					int msg_begin, int msg_end, 
-					size_t *rec_size, 
+					off_t *rec_size, 
 					LIME_type lime_type)
 {
   LRL_RecordWriter *rw;
@@ -168,11 +168,11 @@ LRL_RecordWriter *LRL_open_write_record(LRL_FileWriter *fw,
  *
  * \return number of bytes read
  */
-size_t LRL_read_bytes(LRL_RecordReader *rr, char *buf, 
-		      size_t nbytes)
+off_t LRL_read_bytes(LRL_RecordReader *rr, char *buf, 
+		      off_t nbytes)
 {
   int status;
-  size_t nbyt = nbytes;
+  off_t nbyt = nbytes;
   char myname[] = "LRL_read_bytes";
 
   if (rr == NULL)
@@ -198,11 +198,11 @@ size_t LRL_read_bytes(LRL_RecordReader *rr, char *buf,
  *
  * \return number of bytes written
  */
-size_t LRL_write_bytes(LRL_RecordWriter *rw, char *buf, 
-		       size_t nbytes)
+off_t LRL_write_bytes(LRL_RecordWriter *rw, char *buf, 
+		       off_t nbytes)
 {
   int status;
-  size_t nbyt = nbytes;
+  off_t nbyt = nbytes;
 
   if (rw == NULL)
     return 0;
